refactor: move github queries and commands to github workflow abstractions
