+++ Source/Application/Kysect.Shreks.Application.Commands/Kysect.Shreks.Application.Commands.csproj
@@ -16,6 +16,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Kysect.Shreks.Application.Abstractions\Kysect.Shreks.Application.Abstractions.csproj" />
+      <ProjectReference Include="..\Kysect.Shreks.Application.GithubWorkflow.Abstractions\Kysect.Shreks.Application.GithubWorkflow.Abstractions.csproj" />
     </ItemGroup>
 
 </Project>
