+++ Source/Application/Kysect.Shreks.Application.GithubWorkflow.Abstractions/Commands/CreateOrUpdateGithubSubmission.cs
@@ -2,7 +2,7 @@
 using Kysect.Shreks.Application.Dto.Study;
 using MediatR;
 
-namespace Kysect.Shreks.Application.Abstractions.Github.Commands;
+namespace Kysect.Shreks.Application.GithubWorkflow.Abstractions.Commands;
 
 public static class CreateOrUpdateGithubSubmission
 {
