+++ Source/Application/Kysect.Shreks.Application.GithubWorkflow.Abstractions/Commands/SyncGithubAdminWithMentors.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace Kysect.Shreks.Application.Abstractions.Github.Commands;
+namespace Kysect.Shreks.Application.GithubWorkflow.Abstractions.Commands;
 
 public static class SyncGithubAdminWithMentors
 {
