+++ Source/Application/Kysect.Shreks.Application.GithubWorkflow.Abstractions/Queries/GetCurrentUnratedSubmissionByPrNumber.cs
@@ -2,11 +2,11 @@ using Kysect.Shreks.Application.Dto.Github;
 using Kysect.Shreks.Application.Dto.Study;
 using MediatR;
 
-namespace Kysect.Shreks.Application.Abstractions.Github.Queries;
+namespace Kysect.Shreks.Application.GithubWorkflow.Abstractions.Queries;
 
-public static class GetLastSubmissionByPr
+public static class GetCurrentUnratedSubmissionByPrNumber
 {
     public record Query(GithubPullRequestDescriptor PullRequestDescriptor) : IRequest<Response>;
 
-    public record Response(SubmissionDto Submission);
+    public record Response(SubmissionDto SubmissionDto);
 }
