+++ Source/Application/Kysect.Shreks.Application.GithubWorkflow.Abstractions/Queries/GetSubjectCourseByOrganization.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace Kysect.Shreks.Application.Abstractions.Github.Queries;
+namespace Kysect.Shreks.Application.GithubWorkflow.Abstractions.Queries;
 
 public static class GetSubjectCourseByOrganization
 {
