+++ Source/Application/Kysect.Shreks.Application.GithubWorkflow.Abstractions/Queries/GetUserByGithubUsername.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace Kysect.Shreks.Application.Abstractions.Github.Queries;
+namespace Kysect.Shreks.Application.GithubWorkflow.Abstractions.Queries;
 
 public static class GetUserByGithubUsername
 {
