+++ Source/Application/Kysect.Shreks.Application.Handlers/Github/AddGithubUserAssociationHandler.cs
@@ -4,7 +4,7 @@ using Kysect.Shreks.Core.Users;
 using Kysect.Shreks.DataAccess.Abstractions;
 using Kysect.Shreks.DataAccess.Abstractions.Extensions;
 using MediatR;
-using static Kysect.Shreks.Application.Abstractions.Github.Commands.AddGithubUserAssociation;
+using static Kysect.Shreks.Application.GithubWorkflow.Abstractions.Commands.AddGithubUserAssociation;
 
 namespace Kysect.Shreks.Application.Handlers.Github;
