+++ Source/Application/Kysect.Shreks.Application.Handlers/Github/GetSubjectCourseByOrganizationHandler.cs
@@ -3,7 +3,7 @@ using Kysect.Shreks.Core.SubjectCourseAssociations;
 using Kysect.Shreks.DataAccess.Abstractions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using static Kysect.Shreks.Application.Abstractions.Github.Queries.GetSubjectCourseByOrganization;
+using static Kysect.Shreks.Application.GithubWorkflow.Abstractions.Queries.GetSubjectCourseByOrganization;
 namespace Kysect.Shreks.Application.Handlers.Github;
 
 public class GetSubjectCourseByOrganizationHandler : IRequestHandler<Query, Response>
