+++ Source/Application/Kysect.Shreks.Application.Handlers/Github/GetSubjectCourseGithubAssociationsHandler.cs
@@ -5,7 +5,7 @@ using Kysect.Shreks.Core.SubjectCourseAssociations;
 using Kysect.Shreks.DataAccess.Abstractions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using static Kysect.Shreks.Application.Abstractions.Github.Queries.GetSubjectCourseGithubAssociations;
+using static Kysect.Shreks.Application.GithubWorkflow.Abstractions.Queries.GetSubjectCourseGithubAssociations;
 
 namespace Kysect.Shreks.Application.Handlers.Github;
