+++ Source/Application/Kysect.Shreks.Application.Handlers/Github/RemoveGithubUserAssociationHandler.cs
@@ -3,7 +3,7 @@ using Kysect.Shreks.Core.Users;
 using Kysect.Shreks.DataAccess.Abstractions;
 using Kysect.Shreks.DataAccess.Abstractions.Extensions;
 using MediatR;
-using static Kysect.Shreks.Application.Abstractions.Github.Commands.RemoveGithubUserAssociation;
+using static Kysect.Shreks.Application.GithubWorkflow.Abstractions.Commands.RemoveGithubUserAssociation;
 
 namespace Kysect.Shreks.Application.Handlers.Github;
