+++ Source/Application/Kysect.Shreks.Application.Handlers/Github/SyncGithubAdminWithMentorsHandler.cs
@@ -1,5 +1,5 @@
-﻿using Kysect.Shreks.Application.Abstractions.Github;
-using Kysect.Shreks.Application.Abstractions.Github.Commands;
+﻿using Kysect.Shreks.Application.GithubWorkflow.Abstractions;
+using Kysect.Shreks.Application.GithubWorkflow.Abstractions.Commands;
 using Kysect.Shreks.Common.Exceptions;
 using Kysect.Shreks.Core.SubjectCourseAssociations;
 using Kysect.Shreks.Core.UserAssociations;
