+++ Source/Application/Kysect.Shreks.Application.Handlers/Students/GetGithubUsernamesForSubjectCourseHandler.cs
@@ -3,7 +3,7 @@ using Kysect.Shreks.Core.Specifications.Github;
 using Kysect.Shreks.DataAccess.Abstractions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using static Kysect.Shreks.Application.Abstractions.Github.Queries.GetGithubUsernamesForSubjectCourse;
+using static Kysect.Shreks.Application.GithubWorkflow.Abstractions.Queries.GetGithubUsernamesForSubjectCourse;
 
 namespace Kysect.Shreks.Application.Handlers.Students;
