+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Entities/OrganizationDetailsProvider.cs
@@ -1,4 +1,4 @@
-﻿using Kysect.Shreks.Application.Abstractions.Github;
+﻿using Kysect.Shreks.Application.GithubWorkflow.Abstractions;
 using Kysect.Shreks.Integration.Github.Client;
 using Octokit;
