+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Invites/SubjectCourseGithubOrganizationRepositoryManager.cs
@@ -1,4 +1,5 @@
-﻿using Kysect.Shreks.Integration.Github.Client;
+﻿using Kysect.Shreks.Application.GithubWorkflow.Abstractions;
+using Kysect.Shreks.Integration.Github.Client;
 using Octokit;
 
 namespace Kysect.Shreks.Integration.Github.Invites;
