+++ Source/Kysect.Shreks.WebApi/Controllers/StudentController.cs
@@ -1,6 +1,6 @@
-﻿using Kysect.Shreks.Application.Abstractions.Github.Commands;
-using Kysect.Shreks.Application.Abstractions.Students;
+﻿using Kysect.Shreks.Application.Abstractions.Students;
 using Kysect.Shreks.Application.Dto.Users;
+using Kysect.Shreks.Application.GithubWorkflow.Abstractions.Commands;
 using Kysect.Shreks.Application.Handlers.Students;
 using Kysect.Shreks.Identity.Entities;
 using MediatR;
