+++ Source/Kysect.Shreks.WebApi/Controllers/SubjectCourseController.cs
@@ -1,7 +1,7 @@
-﻿using Kysect.Shreks.Application.Abstractions.Github.Commands;
-using Kysect.Shreks.Application.Abstractions.Study.Commands;
+﻿using Kysect.Shreks.Application.Abstractions.Study.Commands;
 using Kysect.Shreks.Application.Abstractions.Study.Queries;
 using Kysect.Shreks.Application.Dto.Study;
+using Kysect.Shreks.Application.GithubWorkflow.Abstractions.Commands;
 using Kysect.Shreks.Identity.Entities;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
